+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - FF-1429 - Updated FunFair.CodeAnalysis to 5.2.2.830
+- FF-1429 - Updated SonarAnalyzer.CSharp to 8.26.0.34506
 ### Removed
 ### Deployment Changes