+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - FF-1429 - Updated coverlet to 3.1.0
+- FF-1429 - Updated FunFair.CodeAnalysis to 5.2.4.854
 ### Removed
 ### Deployment Changes