+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - FF-1429 - Updated Microsoft.NET.Test.Sdk to 16.11.0
+- FF-1429 - Updated NuGet to 5.11.0
 ### Removed
 ### Deployment Changes